# [MTE] [NFC] use vector to collect globals to tag (#120283)